Moved getColor from ClientHelper to avoid client classload crash on server
