+++ src/main/java/com/sammy/malum/common/effect/CorruptedAerialAura.java
@@ -1,6 +1,6 @@
 package com.sammy.malum.common.effect;
 
-import com.sammy.malum.core.helper.ClientHelper;
+import com.sammy.malum.core.helper.ColorHelper;
 import com.sammy.malum.core.registry.content.SpiritTypeRegistry;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
@@ -12,7 +12,7 @@ public class CorruptedAerialAura extends MobEffect
 {
     public CorruptedAerialAura()
     {
-        super(MobEffectCategory.BENEFICIAL, ClientHelper.getColor(SpiritTypeRegistry.AERIAL_SPIRIT_COLOR));
+        super(MobEffectCategory.BENEFICIAL, ColorHelper.getColor(SpiritTypeRegistry.AERIAL_SPIRIT_COLOR));
         addAttributeModifier(ForgeMod.ENTITY_GRAVITY.get(), "e2306a3e-4ffc-45dc-b9c6-30acb18efab3", -0.40f, AttributeModifier.Operation.MULTIPLY_TOTAL);
     }
