+++ src/main/java/com/sammy/malum/common/effect/SacredAura.java
@@ -1,6 +1,6 @@
 package com.sammy.malum.common.effect;
 
-import com.sammy.malum.core.helper.ClientHelper;
+import com.sammy.malum.core.helper.ColorHelper;
 import com.sammy.malum.core.registry.content.SpiritTypeRegistry;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectCategory;
@@ -10,7 +10,7 @@ public class SacredAura extends MobEffect
 {
     public SacredAura()
     {
-        super(MobEffectCategory.BENEFICIAL, ClientHelper.getColor(SpiritTypeRegistry.SACRED_SPIRIT_COLOR));
+        super(MobEffectCategory.BENEFICIAL, ColorHelper.getColor(SpiritTypeRegistry.SACRED_SPIRIT_COLOR));
     }
 
     @Override
