+++ src/main/java/com/sammy/malum/core/helper/ClientHelper.java
@@ -1,7 +1,6 @@
 package com.sammy.malum.core.helper;
 
 import net.minecraft.client.color.item.ItemColors;
-import net.minecraft.util.FastColor;
 import net.minecraft.world.item.Item;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
@@ -18,17 +17,6 @@ public class ClientHelper
         int g = color.getGreen();
         int b = color.getBlue();
         itemColors.register((stack, i) -> r << 16 | g << 8 | b, item.get());
-    }
-    public static Color getColor(int decimal)
-    {
-        int red = FastColor.ARGB32.red(decimal);
-        int green = FastColor.ARGB32.green(decimal);
-        int blue = FastColor.ARGB32.blue(decimal);
-        return new Color(red, green, blue);
-    }
-    public static int getColor(Color color)
-    {
-        return FastColor.ARGB32.color(color.getAlpha(), color.getRed(), color.getGreen(), color.getBlue());
     }
 
     public static Color darker(Color color, int times) {
